fix: allow sending strings as encrypted replies in echo_client example
the `encrypted_reply` function in the echo client example accepts a Message or a string for the reply parameter.  But supplying a string caused an error because the variable name `reply` was being reused.
This commit introduces a new variable to temporarily hold the reply body, so it can get sent.